Add config generation script and some examples
This commit adds a script "kolla-mesos/cmd/manage.py" to generate
the configuration tree and write it into zookeeper.

Note:
This blatantly copies the config templating in
kolla/anisible/roles/*/templates and all.yml. Maybe some time
in the future this could be shared between the deployments if they don't
diverge too badly.

The method of converting a kolla-ansible service into a kolla-mesos
service is:

 cp ansible/<project>/templates/* ../kolla-mesos/config/<project>/templates/
 cp ansible/<project>/roles/config.yml ../kolla-mesos/config/<project>/<service>_config.yml
 # then edit the above to only include "extra" files.
 cp ansible/<projects>/defaults/main.yml ../kolla-mesos/config/<project>/defaults/main.yml

Co-Authored-By: Michal Rostecki <mrostecki@mirantis.com>

Change-Id: I8ae754989f218ec26e1fd3d77c2a10110c34b90f