Add multiple deployment support
Before we could only deploy one deployment per mesos stack. Now we will be
able to deploy multiple ones. Chnages what have been made:

1) Each time deploy is runned it gen "deployment_id" var. It could be
pre-defined with "--kolla-deployment-id" argument or modified with
"--kolla-deployment-id-prefix" argument. In first case it will be created as
is, in second it will use this prefix and add timestamp to it.

2) Deploy will write ZK configuration to specific node with deployment_id in
it, like "/kolla/my-deploy1", "/kolla/my-deploy2", etc

3) All apps will recive env var with deployment_id and start script will use
it to search for configuration in right ZK node.

4) Configuration of all apps changed so their bootstrap now aware of
deployment_id, so all bootstraping will be uniq per deploy.

5) Each deploy now use deployment_id as a top level group in marathon, so
all deployments a visualy separated in marathon UI.

6) I have to move start script from mariadb node(sic!) to some global node,
for all deployments. It's going to be "/kolla/tools" for now.

How to test(at least):
1) Run deploy with "--kolla-deployment-id-prefix something" - you will get
first deployment with timestamp and prefix.
2) Run deploy with "--kolla-deployment-id somename" - you will get second
deployment with this exact name(warning, second deployment will NOT be
succsessfull in a AIO scenario for obvios reasons)
3) Run deploy without any arguments - you will get third deployment with name
as a timestamp.
4) Check for ZK nodes with kolla-mesos-config

Implements: blueprint multiple-deployment-support
Change-Id: Idbe8bc34b31235c306a6cdf2274f0098c5d5bcd9