Script to show status of kolla-mesos tasks
This script will scan all the kolla-mesos config files to
gather all the registration and requirement variables. It then
reads through the zookeeper store to check for the done status.
It then outputs a table showing all the tasks and their current
set of dependencies and states, making it easy to see when a
task is hung up.

Change-Id: I103e9afad629e6682dbe5c49098fe0b1070e2fd0