Handle unexpected SIGTERM in keystone start_extend.sh
Closes-Bug: 1534180

During a deploy, keystone_setup (running extend_start) was
never finishing. The parent died because it received a SIGINT
signal right after the httpd kill which causes it to exit.

This change adds a trap handler to catch that signal and exit
gracefully. It also adds some additional logging output so it
that the keystone setup process is more visible in the logs.

Change-Id: I1dd120c8c028c1bbfe7a1b6897f7426540ef4c15