Don't use ansible_connection=local to register frameworks properly
Mesos frameworks register to the Mesos cluster using their
hostnames or IP addresses specified in their environment variables.
To get the Mesos UI working in web browsers, it have to be
aware of the exact hostnames accessible outside, not "localhost".
The only way to provide this hostname via inventory_hostname
is to not use local connection.

TrivialFix

Change-Id: Iebff82733a205ff3decb3769cd2b4cfe4385eac9