[WIP] Add state info to commands
Related-Bug: 1519134

Debugging failed deployments can be difficult. Adding states
to the commands in zookeeper makes it easier to see the current
state of affairs in a deployment.

Changes:
- in zookeeper, replace .Done node with node data that contains the
state of a command
- states are: None, running, waiting, error, retry, done
- all commands now have registrations. This allows us to see the state
of all the commands (including the daemon launches)
- updates to the show_tasks script to use the new state info
- some additional logging

Change-Id: Ie86da7a397dc333c4b0ee396eb20c7ff6a7a04a3
TODO: update tests to support this, they are currently failing