Split glance group into 2 groups
Right now both glance services placed in the same "glance" group. But if we
only need glace-api IP Addrs from this group, we cant know which ip is
glance-api and which is glance-registry. So we need to split them.

There is no need to fix group name in other services, since all of them
already looking into 'glance-api' group, so we fixing 2 errors here.

TrivialFix

Change-Id: Ic9c3b2cb6a251d9198afe9362ccdf32a093259aa