Remove ovs chronos job. Move all bootstraps to vswitchd container
We had to run this chronos job on each node(for multinode) and it's was a
bad idea and hard to implement. So I get rid of a chronos job and move all to
vswitchd contaner.

Since check_ovs_db command depends on openvswitch-db/daemon, we cant use
openvswitch-db container for this bootstraps, so I have to move
ovs_ensure_configured.sh to vswitchd.

Partially-Implements: multinode
Change-Id: Icde18f57d6485df30aa27a15762a9e011aa449fb