Set correct NIC number for hostmanager IP resolver
In the Vagrantfile, when using VirtualBox, the IP used for the
hostmanager plugin is retrieved from the wrong network adapter. This
causes the hostmanager to insert the wrong IP address in /etc/hosts. The
NIC numbering starts from 0, and the one used in the code is 2. That
corresponds to the 3rd NIC, which is actually used as the interface to
the Neutron external network. This commit sets the NIC number to 1,
which represents the internal network used by OpenStack.

Change-Id: I87f2b081c57c8eb3109e8d98b75ee438dd0d4b82
Closes-Bug: #1563367
Signed-off-by: Andrei-Lucian Șerb <lucian.serb@icloud.com>